leak_lock_ai/requirements.txt
# FastAPI dependencies
fastapi==0.61.1
uvicorn==0.33.0
pydantic==1.6.1

# .env support
python-dotenv==0.14.0

# packages
openai